refactor: move apollo client code to the components folder
