+++ apps/sports-helsinki/src/domain/place/utils.ts
@@ -1,12 +1,11 @@
 import type { Maybe, PlaceDetailsQuery } from 'events-helsinki-components';
 import { PlaceDetailsDocument } from 'events-helsinki-components';
-import { createApolloClient } from '../clients/eventsFederationApolloClient';
+import { sportsApolloClient } from '../clients/sportsApolloClient';
 
 export const getPlaceDetailsFromCache = (
   id: string
 ): Maybe<PlaceDetailsQuery> => {
-  const eventsApolloClient = createApolloClient();
-  return eventsApolloClient.readQuery<PlaceDetailsQuery>({
+  return sportsApolloClient.readQuery<PlaceDetailsQuery>({
     query: PlaceDetailsDocument,
     variables: { id },
   });
