+++ apps/sports-helsinki/src/pages/articles/[...slug].tsx
@@ -37,7 +37,7 @@ import { ArticleDocument } from 'react-helsinki-headless-cms/apollo';
 import AppConfig from '../../domain/app/AppConfig';
 import getSportsStaticProps from '../../domain/app/getSportsStaticProps';
 import cmsHelper from '../../domain/app/headlessCmsHelper';
-import { apolloClient } from '../../domain/clients/eventsFederationApolloClient';
+import { sportsApolloClient } from '../../domain/clients/sportsApolloClient';
 import serverSideTranslationsWithCommon from '../../domain/i18n/serverSideTranslationsWithCommon';
 
 const NextCmsArticle: NextPage<{
@@ -86,7 +86,7 @@ const NextCmsArticle: NextPage<{
 export async function getStaticPaths() {
   // NOTE: It might not be a good thing to use ApolloClient here,
   // since then the build process depends on external service.
-  const articlePageInfos = await getAllArticles(apolloClient);
+  const articlePageInfos = await getAllArticles(sportsApolloClient);
   const paths = articlePageInfos.map((pageInfo) => ({
     params: { slug: cmsHelper.getSlugFromUri(pageInfo.uri) },
     locale: pageInfo.locale,
@@ -155,7 +155,7 @@ export async function getStaticProps(context: GetStaticPropsContext) {
 
 const getProps = async (context: GetStaticPropsContext) => {
   const language = getLanguageOrDefault(context.locale);
-  const { data: articleData } = await apolloClient.query<
+  const { data: articleData } = await sportsApolloClient.query<
     ArticleQuery,
     ArticleQueryVariables
   >({
@@ -171,7 +171,7 @@ const getProps = async (context: GetStaticPropsContext) => {
   // TODO: Breadcrumbs are unstyled, so left disabled
   const breadcrumbs: Breadcrumb[] = []; // await _getBreadcrumbs(cmsClient, currentArticle);
 
-  return { currentArticle, breadcrumbs, apolloClient };
+  return { currentArticle, breadcrumbs, apolloClient: sportsApolloClient };
 };
 
 /**
