+++ packages/components/src/apollo/__tests__/utils.test.ts
@@ -1,0 +1,20 @@
+import EventsFederationApolloClient from '../EventsFederationApolloClient';
+
+describe('client utils', () => {
+  describe('createInstance', () => {
+    it('should mutate references used for caching the client', () => {
+      // @ts-ignore
+      const mutableCache = new MutableReference();
+
+      // @ts-ignore
+      EventsFederationApolloClient.createInstance({
+        // @ts-ignore
+        mutableCachedClient: mutableCache,
+        // @ts-ignore
+        createClient: () => 1,
+      });
+
+      expect(mutableCache.reference).toStrictEqual(1);
+    });
+  });
+});
