+++ packages/components/src/apollo/index.ts
@@ -1,0 +1,2 @@
+export * from './utils';
+export { default as EventsFederationApolloClient } from './EventsFederationApolloClient';
