+++ packages/components/src/index.ts
@@ -1,4 +1,4 @@
-export * from './apollo/utils';
+export * from './apollo';
 export { AsyncMessage } from './async-message';
 export * from './components';
 export * from './constants';
